Adding meta-tag to enforce standards mode. Fixes #139
Note that we're using IE=edge which is actually discouraged by http://www.alistapart.com/articles/beyonddoctype. We're not too concerned because we plan to test with newer browsers before they're out and fix the actual issues rather than lock our site into the rendering mode for a specific version of the browser. At least that's the plan. :)