Changed the runtime to allow & characters in the request path. Fixes #218
By default, the following characters are not allowed in the request path:

    <,>,*,%,&,:,\,?

Unfortunately, the old gallery allowed any characters in the user name.
It turns out though, that only one existing user account (at least in the
preview site) uses an invalid character, "Patterns & Practices". So I changed
the invalid character list to be the same as the default, but removed just the
&.