UpdateDeployScript
Updated the deployment script as per the structural changes in our
storage container.
Changes to vsmdi to add newly added tests.
Checking in the latest nuget.exe and nuget.core. In team city this gets
picked up from the client build and in local test runs, the setupscript
will pick the latest from our internal nightly drop. But just adding the
nuget.exe and nuget.core to the repo to be on safer side to avoid build
failures.