Preserve whitespace in release notes
Previous version of this code doesn't preserve spaces in release notes.
I understand that you might want to do this a different way, but I
wanted to start a conversation on this.

Without this change, it makes displaying release notes difficult - see
https://nuget.org/packages/Glimpse/1.3.0. In this example I was trying
to nest some of the items - see
http://getglimpse.com/version/install?Glimpse=1.2.0..1.3.0

Also can we please have the line breaks as br's instead of p's. Again
the formatting looks kinda sucky if we are trying to do anything more
than a block of text with release notes.