Migrations add a default row to GallerySettings
Without this, the GallerySettings row cannot be updated when the statistics job is run, and the last updated operation is not stored. This means the download count of each package is incremented by the total count of download operations every 5 minutes, rather than a running total of new operations.

Fixes #1110