In addition to queueing the log messages in memory, also writing it to a local log file. In the unlikely event of the process crashing in a manner that the log messages in memory are not uploaded to blob storage, say, due to critical exceptions such as OOM; log messages can still be found in the local log file created temporarily
The log messages are always written to the local log file synchronously
in addition to writing it in memory. After every 100 or so log messages,
the messages in memory are uploaded to Azure. Once the upload is successful,
the local log file is deleted. Similar to the azure log files which are
sequenced, the local log files are also sequenced. Only difference is once
the ith batch log file is uploaded to azure, ith batch local log file is
deleted from the local filesystem

The local log files are always generated in LocalAppData