Heartbeat emails
An exe to monitor the log files that get written to storage and send
Info, Summary or Alert emails based on the Job health. Alert emails will
be sent if the job is not running or not restarted whereas summary/ info
mails are sent once a day. For a week I'll test this by making this send
mails only to me and then hook this up to pagerduty alerting system