Update Enable-LocalTestMe.ps1 to find makecert
Following the current readme, makecert is located in a Visual Studio extension folder installed by Azure SDK, instead of the Windows SDK that the script previously scanned the registry for.  The scan is still there, but if the scan doesn't find the makecert properly, then the extension folder is tested instead.