Added generic Service Bus serializer (#70)
Added `BrokedMessageSerializer`, a JSON serializer that enforces versioned schemas. Updated the `PackageValidationMessageData` serializer to use this new serializer. This will be used by `PackageSigningValidator` and `PackageCertificatesValidator` to queue messages to their downstream validation jobs.

NOTE: The `BrokedMessageSerializer` does not gracefully handle messages with a different version that what is expected. This will have to be added in later.