[Package Signing] Add Trusted Timestamps to Validation Context (#75)
Each signature on a package is stamped with a signed timestamp from a trusted source. This will be represented by the `TrustedTimestamp` entity. Each `PackageSignature` has one `TrustedTimestamp`. Each `TrustedTimestamp` was signed with one `Certificate`. A `Certificate` may be depended on 0 or more `PackageSignature`s as well as 0 or more `TrustedTimestamp`s.

Miscelleanous changes piggybacking this pull request:
* Added an `Unknown` value to the `PackageSignatureStatus` enum. This will be necessary to distinguish how an invalidated package should be handled (specifically, should we alert on this failed signature validation?). I updated the spec to reflect this new enum value.
* Enforced a default value for the `PackageSignature`'s `CreatedAt` property.