[Package Signing] Fix idempotency (#267)
Previously, the validate package signature job was designed such that it would fail validation requests for  packages that had already been validated unless the validation request was from a revalidation gesture. The idea was that this would make the job "resilient" to unintended state changes caused by message duplication. Essentially, this could happen:

1. Orchestrator sends a `Validate package X's signature` message. Validation passes.
2. Orchestrator resends the `Validate package X's signature` message. Validation fails (even though the signature is still valid).
3. Orchestrator sends a `Revalidate package X's signature` message. Validation passes.

This turned out to be a bad idea as the orchestrator may validate the same package multiple times if it was uploaded multiple times concurrently. With this PR, the new results would be:

1. Orchestrator sends a `Validate package X's signature` message. Validation passes.
2. Orchestrator resends the `Validate package X's signature` message. Validation passes.
3. Orchestrator sends a `Revalidate package X's signature` message. Validation passes.