[Search] Add config for Azure Directory's cache path (#288)
By default, Lucene will store the Azure Directory cache inside of the temp path. On Azure, this is the same drive as the operating system, which only has 20GB of free space. This configuration will allow us to move the Azure Directory cache to a drive with more free space. If `AzureDirectoryCachePath` configuration isn't set, the search service will default to using the temp path.