Read DowloadData from downloads.v1.json (#593)
The goal is to leave the existing download.v1.json data structure Downloads as-is to ensure existing search is unaffected.
Instead of changing Downloads, an API is added to read this file into the more convenient DownloadData class which is flexible enough for comparison and maintaining original case of IDs.
Progress on https://github.com/NuGet/NuGetGallery/issues/6458