Add DownloadDataClient to read and write downloads.v2.json (#591)
This client will be used to read and write the download data that was most recently indexed by Auxiliary2AzureSearch.
This data will be compared to the latest downloads.v1.json and only the changes will be sent to Azure Search.
This strategy is very similar to Owners2AzureSearch.

Progress on https://github.com/NuGet/NuGetGallery/issues/6458