Optimize packageid:{id} version:{version} ignoreFilter=true case (#605)
This case covered the gallery hijack for OData Packages(Id=,Version=).
This accounts for 22.47% of all requests in the past 90 days.
As with the previous optimization PR (#602) switching from a search to a document lookup cut the query time in roughly a half.
Progress on https://github.com/NuGet/NuGetGallery/issues/6466