Extract collector DI and registration client to generic V3 project (#677)
This supports building other collectors that use dependency injection like Catalog2AzureSearch.
Also, the registration client code is not Azure Search specific so a more generic place is better.
This is a PR to move code around that has no impact on behavior.
It allows me to share code between other new collectors and Catalog2AzureSearch without pulling in all of the Azure Search dependencies.