Redo .NET banner (#8967)
The banner is broken on INT as the image wasn't included in the release artifacts. Also the previous implementation used a padding hack to force a container to have a minimal height. This solution did not work well on very wide screens as it caused the background image to be noticeably clipped. This new solution looks much better on very wide screens and removes the padding hack.

Follow up to https://github.com/NuGet/NuGetGallery/pull/8966
Part of: https://github.com/NuGet/Engineering/issues/4226