Reduce default timeout for catalog HTTP requests (#1125)
The old value was 10 minutes. The new value is 1 minute. We saw causes where the job VMs, under high CPU load, would use the entire 10 minutes for the request and retry 2 more times. For this known endpoint, it's better to prefer a lower timeout and force the job to fail if the timeout of 1 minute fails after retries.
Progress on https://github.com/NuGet/Engineering/issues/4767