Drop the message when the entity is hard deleted (#1132)
This is the current behavior for package validation but not symbol validation on the queue-back message. Since a hard delete can easily happen for symbol packages when the symbol package is replaced, this can lead to DLQ messages when the user replaces the symbols soon after publishing the previous version. There were 4 DLQ messages due to it this week.

Resolve https://github.com/NuGet/Engineering/issues/2602