Disable dependabot.yml
There are risks that we discovered associated with enabled Dependabot that we don't want to take now:
After configuring it on Gallery, we had a lot of build breaks and runtime errors. There are also risks of performance regressions.

1ES is enabling 1ES Dependabot (see thread "1ES Dependabot is coming to your repositories for NuGet Service") which should take care of our CG alerts automatically for us.

Decision:
We are postponing enabling Dependabot for when we have more capacity. 
We will do it in priority order, updating some packages first, starting with the ones that are likely to be in security asks (Azure SDKs or anything else that we interact with resources, scary dependencies: SQL, KeyVault, etc.).