Improve ApiKeyV4 based on review feedback
These changes are non-breaking and should produce API keys with similar characteristics.
The ToUpper change is because this code should not be culture aware, and the character set is just base32 anyway.
Added a comment about no entropy impact in Normalize.
Use crypto random instead of Guid.NewGuid for 16 random bytes.