[OIDC] Find matching federated credential policy (policy evaluation) (#10269)
This is a big chunk of the OIDC feature's business logic. A new IFederatedCredentialEvaluator is added which takes a list of federated credential policies and a bearer token and evaluates which trust policy, if any matches the given bearer token.

This is a read-only operation. The creation of the short-lived API key will be performed at a higher level.