Add a polyfill to enable newer compiler features (#10315)
New versions of C# offer many features that improve quality of life:
nullable annotations, record types, init and required properties, etc.
These can be enabled in projects targeting older framework versions by
adding the right attributes and types to the compilation.

PolySharp is a source generator which automatically does this.  For each
project, it will generate all of the types/attributes which are missing.
In the case of an InternalsVisibleTo reference, it is smart enough to
only define types in the second assembly that are not being consumed
from the first assembly (avoiding duplicate type definitions).