Delete unused load test suite (#10651)
These have not run in ages (couldn't find how long, likely over a year). The test coverage they provide is duplicated in other tests and they are not even executed as load tests. Even if they were, they would not sufficiently cover our performance hot spots and are not run in a scientific way.